requirements.txt
bcrypt==3.2.2
cffi==1.15.1
cryptography==37.0.4
distlib==0.3.4
filelock==3.7.1
hurry.filesize==0.9
libtorrent==2.0.6
numpy==1.22.4
opencv-python==4.6.0.66
paramiko==2.11.0
platformdirs==2.5.2
pycparser==2.21
PyNaCl==1.5.0
PyQt3D==5.15.5
PyQt5==5.15.7
PyQt5-sip==12.11.0
PyQt6==6.3.1
PyQt6-Qt6==6.3.1
PyQt6-sip==13.4.0
PyQtChart==5.15.6
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
PyQtWebEngine==5.15.6
scp==0.14.4
six==1.16.0
virtualenv==20.14.1
